Tests: More ordering for friendliness
