+++ StackExchange.Redis.Tests/MultiMaster.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.IO;
+﻿using System.IO;
 using System.Linq;
 using System.Net;
 using Xunit;
